refactor(agent): extract shared sync canonicalization helpers into sync_canonical
Behaviour-neutral extraction ahead of the software_licensing daily-sync source
(ADR-0024): sanitize_utf8_strict, clamp_field, and sha256_hex move from
sync_source_installed_software.cpp (and the verbatim duplicate in
sync_source_device_ci.cpp) into one shared TU, agents/core/src/sync_canonical.
clamp_field now takes the per-source field cap explicitly; each source keeps
its comment-coordinated cap constant.

The server-side copies (utf8_sanitize.hpp and the per-seam clamps) are
untouched and remain comment-coordinated: the agent/server byte-identity
contract is unchanged, proven by the existing clean and dirty cross-pin hash
tests on both sources (agent suite: [sync] 23 cases / [device_ci] 5 cases, all
green after the extraction).

sha256_hex's declaration moves from sync_source_installed_software.hpp to
sync_canonical.hpp; app_perf/device_ci includes repointed.
