+++ agents/core/src/sync_source_installed_software.cpp
@@ -1,8 +1,8 @@
 #include "sync_source_installed_software.hpp"
 
 #include "local_dispatcher.hpp"
+#include "sync_canonical.hpp" // sanitize_utf8_strict / clamp_field / sha256_hex
 
-#include <openssl/evp.h>
 #include <spdlog/spdlog.h>
 
 #include <algorithm>
@@ -34,99 +34,12 @@ constexpr std::size_t kMaxFieldLen = 1024;
 // installed software.
 constexpr std::size_t kMaxBlobBytes = 3u * 1024 * 1024;
 
-// Replace every byte that is not part of a valid UTF-8 sequence with U+FFFD (the
-// replacement character, EF BF BD). "Valid" is exactly what PostgreSQL's UTF8
-// server-encoding accepts: no overlong forms, no surrogates (U+D800..U+DFFF),
-// nothing above U+10FFFF. A field that reaches PG with an invalid byte is
-// rejected with SQLSTATE 22021, which rolls back the whole full-replace
-// transaction and makes the agent resend the identical poison forever
-// (governance UP-IN1). The installed_apps plugin already sanitises its output, so
-// this is defence-in-depth against any future SyncSource (and a plugin scrub that
-// is structurally laxer than PG, e.g. one that passes surrogate/overlong forms).
-//
-// This MUST run BEFORE the length clamp (U+FFFD is 3 bytes, so scrubbing can grow
-// the field; clamping first would apply a different 1024-byte budget on each
-// side) and MUST be byte-for-byte identical to the server's copy in
-// inventory_ingestion.cpp (parse_software_blob), or the agent's and server's
-// canonical hashes diverge → permanent always-full.
-std::string sanitize_utf8_strict(std::string_view s) {
-    std::string out;
-    out.reserve(s.size());
-    const std::size_t n = s.size();
-    const auto cont = [&](std::size_t j) -> bool {
-        return j < n && (static_cast<unsigned char>(s[j]) & 0xC0) == 0x80;
-    };
-    std::size_t i = 0;
-    while (i < n) {
-        const unsigned char c = static_cast<unsigned char>(s[i]);
-        const unsigned char c1 = i + 1 < n ? static_cast<unsigned char>(s[i + 1]) : 0;
-        std::size_t len = 0;
-        bool ok = false;
-        if (c < 0x80) {
-            ok = true;
-            len = 1;
-        } else if (c >= 0xC2 && c <= 0xDF) {
-            ok = cont(i + 1);
-            len = 2;
-        } else if (c == 0xE0) {
-            ok = c1 >= 0xA0 && c1 <= 0xBF && cont(i + 2); // reject overlong 80..9F
-            len = 3;
-        } else if (c >= 0xE1 && c <= 0xEC) {
-            ok = cont(i + 1) && cont(i + 2);
-            len = 3;
-        } else if (c == 0xED) {
-            ok = c1 >= 0x80 && c1 <= 0x9F && cont(i + 2); // reject surrogates A0..BF
-            len = 3;
-        } else if (c >= 0xEE && c <= 0xEF) {
-            ok = cont(i + 1) && cont(i + 2);
-            len = 3;
-        } else if (c == 0xF0) {
-            ok = c1 >= 0x90 && c1 <= 0xBF && cont(i + 2) && cont(i + 3); // reject overlong 80..8F
-            len = 4;
-        } else if (c >= 0xF1 && c <= 0xF3) {
-            ok = cont(i + 1) && cont(i + 2) && cont(i + 3);
-            len = 4;
-        } else if (c == 0xF4) {
-            ok = c1 >= 0x80 && c1 <= 0x8F && cont(i + 2) && cont(i + 3); // reject > U+10FFFF
-            len = 4;
-        }
-        if (ok) {
-            out.append(s.data() + i, len);
-            i += len;
-        } else {
-            out.append("\xEF\xBF\xBD", 3); // U+FFFD, advance one byte
-            i += 1;
-        }
-    }
-    return out;
-}
-
-std::string clamp_field(std::string_view raw) {
-    // 1. Scrub to valid UTF-8 FIRST (see sanitize_utf8_strict — order is
-    //    load-bearing for the agent/server hash coordination).
-    std::string f = sanitize_utf8_strict(raw);
-    // 2. Truncate on a UTF-8 codepoint boundary, never mid-sequence. After the
-    //    scrub the field is valid UTF-8, so backing up over trailing continuation
-    //    bytes (0b10xxxxxx) lands on a codepoint start. MUST match the server's
-    //    parse_software_blob field clamp (inventory_ingestion.cpp) byte-for-byte.
-    if (f.size() > kMaxFieldLen) {
-        std::size_t end = kMaxFieldLen;
-        while (end > 0 && (static_cast<unsigned char>(f[end]) & 0xC0) == 0x80)
-            --end;
-        f.resize(end);
-    }
-    // 3. Strip the canonical framing separators (and NUL) so a value can never
-    //    corrupt the wire blob's structure (0x1F field / 0x1E record separators) —
-    //    the server splits on exactly those bytes. U+FFFD contains none of them.
-    std::string out;
-    out.reserve(f.size());
-    for (char c : f) {
-        if (c == '\x1f' || c == '\x1e' || c == '\0')
-            continue;
-        out.push_back(c);
-    }
-    return out;
-}
+// The UTF-8 scrub + field clamp (sanitize_utf8_strict / clamp_field) live in
+// sync_canonical.{hpp,cpp} — one agent-side implementation shared by every
+// daily-sync source. The server's copy in inventory_ingestion.cpp
+// (parse_software_blob) must stay byte-for-byte identical, or the agent's and
+// server's canonical hashes diverge → permanent always-full. This source's
+// field cap is kMaxFieldLen above (comment-coordinated with the server seam).
 
 // Sort/dedup key walks ALL 12 v2 fields in blob order. MUST mirror the server's
 // entry_less/entry_equal (software_inventory_store.cpp) or the two sides'
@@ -167,21 +80,6 @@ bool entry_equal(const SwEntry& a, const SwEntry& b) {
 
 } // namespace
 
-std::string sha256_hex(const std::string& in) {
-    unsigned char md[EVP_MAX_MD_SIZE];
-    unsigned int len = 0;
-    if (EVP_Digest(in.data(), in.size(), md, &len, EVP_sha256(), nullptr) != 1)
-        return {};
-    static const char* kHex = "0123456789abcdef";
-    std::string out;
-    out.reserve(static_cast<std::size_t>(len) * 2);
-    for (unsigned int i = 0; i < len; ++i) {
-        out.push_back(kHex[md[i] >> 4]);
-        out.push_back(kHex[md[i] & 0x0f]);
-    }
-    return out;
-}
-
 std::vector<SwEntry> parse_installed_apps_output(const std::string& out) {
     std::vector<SwEntry> entries;
     std::size_t pos = 0;
@@ -218,7 +116,7 @@ std::vector<SwEntry> parse_installed_apps_output(const std::string& out) {
         // Blob contract v2 field order; missing trailing tokens → empty fields.
         SwEntry e;
         const auto field = [&tok](std::size_t i) -> std::string {
-            return tok.size() > i ? clamp_field(tok[i]) : std::string{};
+            return tok.size() > i ? clamp_field(tok[i], kMaxFieldLen) : std::string{};
         };
         e.name = field(1);
         e.version = field(2);
